refactoring notifications
