+++ app/src/main/java/com/emikhalets/mydates/foreground/EventsReceiver.kt
@@ -7,11 +7,7 @@ import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.workDataOf
 import com.emikhalets.mydates.utils.AppAlarmManager
-import com.emikhalets.mydates.utils.DATA_NOTIF_DAY
-import com.emikhalets.mydates.utils.DATA_NOTIF_MONTH
-import com.emikhalets.mydates.utils.DATA_NOTIF_TODAY
-import com.emikhalets.mydates.utils.DATA_NOTIF_TWO_DAY
-import com.emikhalets.mydates.utils.DATA_NOTIF_WEEK
+import com.emikhalets.mydates.utils.AppNotificationManager
 import com.emikhalets.mydates.utils.di.appComponent
 import com.emikhalets.mydates.utils.extentions.launchMainScope
 
@@ -28,11 +24,11 @@ class EventsReceiver : BroadcastReceiver() {
                 val isToday = context.appComponent.appPreferences.getNotificationToday()
 
                 val data = workDataOf(
-                    DATA_NOTIF_MONTH to isMonth,
-                    DATA_NOTIF_WEEK to isWeek,
-                    DATA_NOTIF_TWO_DAY to isTwoDays,
-                    DATA_NOTIF_DAY to isDays,
-                    DATA_NOTIF_TODAY to isToday
+                    AppNotificationManager.DATA_NOTIFICATION_MONTH to isMonth,
+                    AppNotificationManager.DATA_NOTIFICATION_WEEK to isWeek,
+                    AppNotificationManager.DATA_NOTIFICATION_TWO_DAY to isTwoDays,
+                    AppNotificationManager.DATA_NOTIFICATION_DAY to isDays,
+                    AppNotificationManager.DATA_NOTIFICATION_TODAY to isToday
                 )
                 val work = OneTimeWorkRequestBuilder<EventWorker>().setInputData(data).build()
                 WorkManager.getInstance(context).enqueue(work)
