+++ app/src/main/java/com/emikhalets/mydates/foreground/UpdateEventsWorker.kt
@@ -6,8 +6,8 @@ import androidx.work.WorkerParameters
 import com.emikhalets.mydates.data.database.AppDatabase
 import com.emikhalets.mydates.data.database.CompleteResult
 import com.emikhalets.mydates.data.repositories.RoomRepositoryImpl
+import com.emikhalets.mydates.utils.AppNotificationManager
 import com.emikhalets.mydates.utils.di.appComponent
-import com.emikhalets.mydates.utils.sendErrorUpdateNotification
 import java.util.*
 
 class UpdateEventsWorker(context: Context, parameters: WorkerParameters) :
@@ -19,7 +19,7 @@ class UpdateEventsWorker(context: Context, parameters: WorkerParameters) :
 
         when (repo.updateEvents()) {
             is CompleteResult.Error -> {
-                sendErrorUpdateNotification(applicationContext)
+                AppNotificationManager.sendErrorUpdateNotification(applicationContext)
             }
             CompleteResult.Complete -> {
                 applicationContext.appComponent.appPreferences.setEventsLastUpdateTime(Date().time)
